requirements.txt
PyGithub==1.55
Mastodon.py==1.5.1